fixed macos compatibiliy issues arising from assuming `size_t`, `uint64_t`, and `VkDeviceSize` are the same type
missed one

properly link shaderc_combined in a platform independent manner

more cmake script fixes

macos compatibiltiy changes

fixed evaluation of ${APPLE} and re-added char8_t flags

+++ include/vuk/runtime/vk/BufferAllocator.hpp
@@ -28,7 +28,7 @@ namespace vuk {
 		DeviceResource* upstream;
 		std::mutex mutex;
 		std::atomic<int> current_buffer = -1;
-		std::atomic<uint64_t> needle = 0;
+		std::atomic<size_t> needle = 0;
 		MemoryUsage mem_usage;
 		BufferUsageFlags usage;
 		// TODO: convert to deque
